requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
httpx
tenacity
python-dotenv